lecture6/classifier/requirements.txt
Flask==1.1.2
Jinja2==2.11.2
Werkzeug==1.0.1
gunicorn==20.0.4
numpy==1.18.1
Pillow==7.1.2
http://download.pytorch.org/whl/cpu/torch-1.3.1%2Bcpu-cp37-cp37m-linux_x86_64.whl
torchvision==0.4.2
torch==1.
